feat: listen to app links in tabview screen
