+++ lib/main.dart
@@ -1,7 +1,3 @@
-import 'dart:async';
-import 'dart:developer';
-
-import 'package:app_links/app_links.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -14,60 +10,12 @@ void main() async {
   runApp(const MyApp());
 }
 
-class MyApp extends StatefulWidget {
+class MyApp extends StatelessWidget {
   const MyApp({super.key});
-
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  final _navigatorKey = GlobalKey<NavigatorState>();
-  late AppLinks _appLinks;
-  StreamSubscription<Uri>? _linkSubscription;
-
-  Future<void> initAppLinks() async {
-    _appLinks = AppLinks();
-
-    // Check initial link if app was in cold state (terminated)
-    final appLink = await _appLinks.getInitialAppLink();
-    if (appLink != null) {
-      log('getInitialAppLink: $appLink');
-      openAppLink(appLink);
-    }
-
-    // Handle link when app is in warm state (front or background)
-    _linkSubscription = _appLinks.uriLinkStream.listen((uri) {
-      log('onAppLink: $uri');
-      openAppLink(uri);
-    });
-  }
-
-  void openAppLink(Uri uri) {
-    log("Open Link Called");
-    log(uri.pathSegments.last);
-    //_navigatorKey.currentState?.pushNamed(AppRoutes.profile);
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      Get.defaultDialog(title: uri.pathSegments.last);
-    });
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    initAppLinks();
-  }
-
-  @override
-  void dispose() {
-    _linkSubscription?.cancel();
-    super.dispose();
-  }
 
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
-      navigatorKey: _navigatorKey,
       title: 'Resonate',
       theme: ThemeData(
           scaffoldBackgroundColor: AppColor.bgBlackColor,
@@ -79,7 +27,7 @@ class _MyAppState extends State<MyApp> {
             errorColor: const Color(0xFDFF0000),
           ),
           fontFamily:
-              GoogleFonts.poppins(fontWeight: FontWeight.w500).fontFamily,
+          GoogleFonts.poppins(fontWeight: FontWeight.w500).fontFamily,
           iconTheme: Get.theme.iconTheme.copyWith(
             color: AppColor.yellowColor,
           ),
